Script to post to Facebook groups
I tested this with a group I am the admin and one where I am not. Worked for both. Simple python script that uses Selenium to post a message to a Facebook group.

All you need to add is your login info and the URL for the group page.

I still consider myself new to programming. Comments, improvements, suggestions, critiques all welcome.

Thanks.